Refs #28920 - Settings resolve permission name
