+++ app/models/concerns/authorizable.rb
@@ -1,5 +1,6 @@
 module Authorizable
   extend ActiveSupport::Concern
+  include PermissionName
 
   def check_permissions_after_save
     return true if Thread.current[:ignore_permission_check]
@@ -30,10 +31,6 @@ module Authorizable
   def authorized?(permission)
     return false if User.current.nil?
     User.current.can?(permission, self)
-  end
-
-  def permission_name(action)
-    self.class.find_permission_name(action)
   end
 
   included do
@@ -106,18 +103,6 @@ module Authorizable
       yield
     ensure
       Thread.current[:ignore_permission_check] = original_value
-    end
-
-    def find_permission_name(action)
-      type = Permission.resource_name(self)
-      permissions = Permission.where(:resource_type => type).where(["#{Permission.table_name}.name LIKE ?", "#{action}_%"])
-
-      # some permissions are grouped for same resource, e.g. edit_comupute_resources and edit_compute_resources_vms, in such case we need to detect the right permission
-      if permissions.size > 1
-        permissions.detect { |p| p.name.end_with?(type.underscore.pluralize) }.try(:name)
-      else
-        permissions.first.try(:name)
-      end
     end
   end
 end
